requirements.txt
amqp==2.6.1
billiard==3.6.4.0
celery==4.4.7
configparser==4.0.2
contextlib2==0.6.0.post1
Django==1.11.29
django-ckeditor==5.6.0
django-js-asset==1.2.3
importlib-metadata==2.1.3
kombu==4.6.11
pathlib2==2.3.7.post1
psycopg2-binary==2.8.6
pytz==2022.6
scandir==1.10.0
six==1.16.0
typing==3.10.0.0
vine==1.3.0
zipp==1.2.0
